refactor: 패키지 구조 정리 및 임포트 경로 수정
1. Color, Project, Task 관련 클래스들을 새로운 패키지 구조로 이동한다.
2. 관련 테스트 클래스에서 수정된 패키지 경로에 맞게 임포트를 변경한다.
3. 불필요한 임포트를 제거하고 코드 일관성을 유지한다.

Resolves #7
