+++ src/main/java/com/todoservice/greencatsoftware/domain/project/entity/Project.java
@@ -3,12 +3,11 @@ package com.todoservice.greencatsoftware.domain.project.entity;
 import com.todoservice.greencatsoftware.common.enums.Status;
 import com.todoservice.greencatsoftware.common.enums.Visibility;
 import com.todoservice.greencatsoftware.common.superEntity.SuperEntity;
-import com.todoservice.greencatsoftware.domain.color.Color;
+import com.todoservice.greencatsoftware.domain.color.entity.Color;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.Setter;
-import org.hibernate.validator.constraints.Length;
 
 @Entity
 @Getter
