+++ src/main/java/com/todoservice/greencatsoftware/domain/project/model/ProjectRepository.java
@@ -1,4 +1,4 @@
-package com.todoservice.greencatsoftware.domain.project;
+package com.todoservice.greencatsoftware.domain.project.model;
 
 import com.todoservice.greencatsoftware.domain.project.entity.Project;
 import org.springframework.data.jpa.repository.JpaRepository;
