+++ src/main/java/com/todoservice/greencatsoftware/domain/task/entity/Task.java
@@ -4,13 +4,12 @@ import com.todoservice.greencatsoftware.common.enums.DayLabel;
 import com.todoservice.greencatsoftware.common.enums.Priority;
 import com.todoservice.greencatsoftware.common.enums.Status;
 import com.todoservice.greencatsoftware.common.superEntity.SuperEntity;
-import com.todoservice.greencatsoftware.domain.color.Color;
+import com.todoservice.greencatsoftware.domain.color.entity.Color;
 import com.todoservice.greencatsoftware.domain.project.entity.Project;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
 import lombok.Getter;
 import lombok.Setter;
-import org.hibernate.validator.constraints.Length;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
