+++ src/test/java/com/todoservice/greencatsoftware/domain/task/TaskRepositoryTest.java
@@ -4,11 +4,12 @@ import com.todoservice.greencatsoftware.common.enums.DayLabel;
 import com.todoservice.greencatsoftware.common.enums.Priority;
 import com.todoservice.greencatsoftware.common.enums.Status;
 import com.todoservice.greencatsoftware.common.enums.Visibility;
-import com.todoservice.greencatsoftware.domain.color.Color;
-import com.todoservice.greencatsoftware.domain.color.ColorRepository;
-import com.todoservice.greencatsoftware.domain.project.ProjectRepository;
+import com.todoservice.greencatsoftware.domain.color.entity.Color;
+import com.todoservice.greencatsoftware.domain.color.model.ColorRepository;
+import com.todoservice.greencatsoftware.domain.project.model.ProjectRepository;
 import com.todoservice.greencatsoftware.domain.project.entity.Project;
 import com.todoservice.greencatsoftware.domain.task.entity.Task;
+import com.todoservice.greencatsoftware.domain.task.model.TaskRepository;
 import jakarta.validation.ConstraintViolationException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
